Se crea el archivo compiler para compilar archivos, se crean las clases directorios para la tabla de variables y el directorio de funciones, Se crea el archivo de prueba
