+++ Geometrik/compiler.py
@@ -1,0 +1,5 @@
+from scannerLex import scanner
+from parserYacc import parse
+
+tokens = scanner("TestCode")
+parse(tokens, "TestCode")
