+++ README.md
@@ -1,2 +1,11 @@
 # Geometrik
 Proyecto de Diseño de Compiladores
+
+Para correr el compilador ejecutar el comando:
+
+    python compiler.py
+
+Es necesario estar ubicado en el directorio donde se encuentra el archivo compiler.py (/Geometrik/Geometrik)
+
+compiler.py prueba un archivo de texto con codigo llamado TestCode, si se desea compilar otro archivo
+es necesario modificar el nombre del archivo a compilar en el archivo compiler.py
